Azure: Remove deployments in existing resource groups
Azure Shoot deployments into existing resource groups were available code wise but are disabled due to issues with resourche leftovers. In general existing resource group deployments were only a prerequisite for deployments into existing virtual networks (vnets) which need to be in the cluster resource group. This has changed as it is meanwhile possible to deploy a Shoot in an existing vnet which is located in another resource group then the cluster resources. Therefore there is no known use case for deployment into existing resource groups.

With the removal of the deployment into existing resource groups option there was also some need to restructure the Azure infrastructure terraform chart.